+++ KoruMsSqlYedek.Win/Properties/AssemblyInfo.cs
@@ -29,5 +29,5 @@ using System.Runtime.InteropServices;
 //      Yapı Numarası
 //      Düzeltme
 //
-[assembly: AssemblyVersion("0.42.6.0")]
-[assembly: AssemblyFileVersion("0.42.6.0")]
+[assembly: AssemblyVersion("0.42.9.0")]
+[assembly: AssemblyFileVersion("0.42.9.0")]